[refactor] 핸들러의 성격에 맞게 클래스 네임을 변경하였습니다.
기존 UserHandler 는 UserApiHandler 로 변경하고
동적으로 렌더링이 필요한 핸들러인 IndexHandler 를 UserHandler 로 변경하였습니다.
